Properly positioning the cursor on a LineWise paste.
Besides fixing parity issues between real Vim and VsVim, this also fixes the problem where a shift-P command would paste three times. The root cause was that my code change in revision 6d83cbb was throwing an exception because of a mismatched ITextSnapshot/SnapshotPoint. This meant that the KeyProcessor didn't mark the key as having been handled, resulting in IVimBuffer.ProcessInput being called (for the same keystroke) in response to KeyProcessor.TextInput and VsCommandFilter.Exec.