Make VsVim properly F5 on a brand new checkout.
There are two tricks here: first, in the absence of a .suo file, Visual Studio uses the first project listed in the .sln file as the default startup project. Second, we can add the startup action to the .csproj file itself. If you change it within VS, the .csproj.user file will still override properly.