Fixes the issue where Tab in Visual Studio 2010 SP1 was causing the wrong line to be indented in several scenarios (most prominently in R# after issuing an 'o' or 'O' command).
The root cause of this is the changed behavior of the IEditorOperations.Indent command.  In SP1 it altered it's behavior to rely on ITextSelection.VirtualSelectedSpans more heavily in RTM.  This revealed a bug in the way that VsVim was positioning the caret.  In the majority of caret movements it was failing to reset the selection (movement from an API and reseting the selection are different actions).  So operations like 'o' would move the caret down a line but leave ITextSelection.VirtualSelectedSpans pointing at the previous line.

Changed the core move API to reset the selection on every move.  Fixed up the few places where a reset was explicitly not wanted (entering Visual Mode for example)

closes #606