Moved the highlight search tagger to be asynchronous.  Previously it was synchronous and causing perf problems in Visual Studio when a large number of files were open and :hlsearch was enabled.
There is data source which all highlight search taggers look to (the
last search pattern).  When that was updated it caused every tagger
instance to raise TagsChanged.  Visual Studio would then go back and ask
every one for GetTags on the entire buffer even if the ITextView wasn't
visible at the time.

This happened on idle.  But because the requests came one after the
other it made it little different than it being completely synchronous.

This change simply makes it asynchronous to work around the problem