This is a partial fix for issue #709.  This resolves the mapping problems which occur whis Shift is applied to non-alpha keys such as Return or Tab: <S-RETURN> or <S-TAB>.
VsVim already supported this style of mapping the problem was a bug in
the translation of input at the Visual Studio level.  Both Tab and
Return end up as Visual Studio commands and neither carries the Shift
modifier with them.

The fix was to have the code which does the mapping consult the keyboard
to see if Shift (or other modifiers) were being held down during the
mapping.

Note: Visual Studio does have this information at the start of input
routing.  You can set a VS shortcut to execute on a Shift+Enter but by
the time it gets to IOleCommandTarget it's dumped this extra
information.

issue #709