This fixes a bug in the repeat infrastructure.  If an 'o' command executed but changed no text and then was immediately followed by another command such as 'cw', 'dw', etc ... the 'o' command would be incorrect linked with the following command.
The bug was how insert mode handled <Esc>.  In most cases the caret is
moved left as a command.  This command would correctly link in with the
edit command such as 'o' and hence register as the end of the command.
When the caret was in virtual space though the caret was directly moved
vs. registering as a command.  This caused the 'o' to remain open and
accepeted the next executed command.

The fix is to always leave insert mode with a command.  Let the
implementation deal with virtual space problems

closes #748