Made \n match \r, \n, and \r\n for both search & replace
To do this I had to use a non-capturing group in the regex, so one caveat
and possible future bug is that you now can't use \n in a character class
(i.e. `/bugs[\n :]are great/` will be actually be the pattern
`/bugs[(?:\r?\n|\r) :]are great/` which isn't the regex you would expect).

The fix for this would actually be a little invasive because you would
have to add some sort of `IsInsideCharacterClass` property to
`VimRegex.Data` for use in `VimRegexFactory.ProcessEscapedChar`. And
honestly, I can't think of a real fix for this (char classes don't
recognize order so `[\r\n :]` would never be meaningful). The best we could
ever do is let `\n` mean a literal `\n` when inside a character class, but
even still it wouldn't ever do what you expect.

This fixes #446 for all cases @MartinLemburg seems to be concerned with
except `^Q` (which seems like it might be a different unimplemented
feature). It also doesn't seem to break any other cases, AFAIK.