Fixes the issue where Normal Mode commands were being processed as direct edits if Intellisense was actively displayed.  This could happen if the user explicitly forced intellisense (CTRL+Space) or it was passively displayed by another extension (MindScape).
For most commands it's straight forward that normal mode should win.
However there was some debate on how say Tab, Enter and the arrow keys
should be handled since they are much more associated with Intellisense

After a bit of debate it was decided that the keys very much associated
with Intellisense should be processed by intellisense.  The rest should
go to normal mode for processing

closes #739