Moved cursor out of virtual space on dedent if line is blank
This fixes #746 by moving the caret to the end of the inserted spaces. The
previous behavior was inserting the correct amount of space, but leaving
the virtual space as well, so the cursor appeared to be extra-indented.

There is an additional change that could be made to keep the cursor at
it's "current" position - as in previous column minus dedent space. This
would be nice just because it mimicks what Vim does. But it's also nice to
not have dedent operations moving your cursor all over the place.