Substitute can use last search again
I removed references to `SubstituteFlags.UsePreviousSearchPattern` from
the RunSubstitute member. It appears that this flag should only be used in
`:&[flags]` to run the same substitute with different flags. We were using
this flag in the actual `:substitute` routine. As far as I can tell, there
isn't a good reason for this.