Fixed caret behavior for `#if DEBUG`
If the caret is over like `#if DE^UG` and you hit `%`, Vim is supposed to
still find the `#if` at the beginning of the line. This commit fixes it
by, on failure to find the token, searching again from the beginning of
the line.

The only catch is that this commit also makes it so that `if { stuff^`
will still find the matching brace. This is non-standard behavior, but
quite honestly it's a corner case that we shouldn't have to worry about,
and probably a little bit convenient.