Further cleaned up the mapping of user input to KeyInput values.  On Croatian keyboards the '>' char when typed was coming across as Shift+> which caused it to not match up with the shift command.  After some experimentation it appears that even though you can create mappings for key combinations like <S->> Vim smoothes them out to simply > at input time.  The mapping exists but isn't hit.
There appears to be no documentation as to what characters are smoothed
out in this fashion.  I did some experimentation and found what appears
to be the critical set.

Also fixed NormalMode so that it will report it can process any letter
(even if it doesn't map to a command) and then ignore it.  International
characters were slipping through and turning into input in the buffer

closes #824