Fixes the repeat of a replace mode edit. There was a bug that caused it to repeat as an insert intstead of a replace.
Core problem is that insert and replace were being represented by the
same command.  During a repeat the mode isn't entered, just the commands
are replayed so overwrite was never set.  Added a new command
DirectReplace, to match DirectInsert, which ensures that the overwrite
option is properly set before doing the edit

closes #864