Fix conflict with multiple JetBrains products installed
The JetBrains set of products reuses at a code level several VS
components.  In particular it reuses the tagging infrastructure between
R#, dotTrace and dotCover.

The VsVim external edit monitoring service was only doing a name match
on the tagger infrastructure.  When mulitple JetBrains products were
installed it would pick the first it saw, which was almost never R#, and
listen to that tagger for rename events.  Since it wasn't the R# tagger
it didn't get the notification and R# support appeared to be broken.

This change alters the code to do a smarter match.   It's smarter but
more restrictive.  If R# changes the assembly name between versions then
it will break VsVim.  Dev10 is in the name of the assembly and it's
unclear if it's referring to the Dev10 editor or specifically Visual
Studio 2010.  I tried peaking at the R# 7.0 support which includes Dev11
but the link is currently down.  Added some code to fall back to a less
restrictive match in case a strong one wasn't found.

closes #635

Conflicts:

	VsVimShared/VsVimHost.cs