Added IVim.GetOrCreateVimBufferForHost
This method is necessary because the MEF component startup order is
undefined.  It just so happens today that Visual Studio will startup the
main host in such a way that it creates the IVimBuffer before the main
ITagger implementations are created.  Hence it's there for them to see.

This is not guaranteed though.  It's very possible for the ITagger
implementations to be created first.  In that case if they used
IVim.GetVimBuffer they wouldn't see an IVimBuffer and wouldn't be
created.

This API is a middle ground.  It lets the IVimHost decide if an
ITextView is elligable for an IVimBuffer and removes the ordering
issues.