Fixed the blank options page
I introduced a regression between 1.2.2 and 1.3.0 which brought back the
empty options page on certain non-English builds.  Chinese and Japanese
in particular.

This came as a result of a typo in a refactoring I did in the
KeyBindingService.cs in the following change

f20d6139340b855ef59621ac6af42f54cd1b0f55

The _importantScopeSet field was initialized with the default scope,
which is actually a fallback, instead of the scopes which are derived
from registry entries.

fixes #476