Fixed the large VSIX issue
The VSIX for VsVim had approach 7MB in size.  This was due to the WPF
assemblies, including Microsoft.VisualStudio.Text.UI.WPF,
PresentationBase, etc ... from being embedded.

It turns out that EditorUtils was referencing
Microsoft.VisualStudio.Text.UI.Wpf but not VsVim.  The dependency code
saw the mismatch, decided it was a private need to be depoyled assembly
and deployed it plus all the WPF dependencies.

Even though VimCore doesn't depend on or use Wpf adding the ref now to
prevent the extremely large VSIX