First step of a larger refactoring
This is the first step of a larger refactoring effort.  This step hid
most of what used to be a set of private types into their own namespace
to facilitate testing.

The KeyUtil type is still an ambient authority.  It needs to be
expressed via a MEF interface.  This is a bit more of an impactful
change though.  Going to focus on testing for a bit and make sure I have
this particular problem nailed before I get into that