Fixed navigation in metadata files
The new keyboard handling code was failing to execute commands in files
that were readonly (namely metadata files).  The immediate problem was
the KeyProcessor implementation wasn't recieving the TextInput event for
key strokes even though it did see the KeyDown and KeyUp events.

The root cause is the VsCodeWindowAdapter class.  It was intercepting
the WM_CHAR message which eventually becomes TextInput in the
PreProcessMessage method.  If it determined the buffer was readonly it
would suppress any messages it deemed to be an edit.

This can be overriden a couple of ways but I chose to do it by emulating
the search component and setting the SearchInProgress flag.   This lets
the commands through that VsVim is interested in.  Another option was
IOleComponent and get the PreProcessMessage earlier (this is how
PowerToys quick find works) but it's too easy to conflict with other
components there (no nesting with IOleComponent, only one can be
tracking at a time).

closes #947