Fix ability to map <C-;>
There is a hidden Visual Studio command that is bound to <C-;>.  There
was a bug that was preventing VsVim from being able to properly unmap
this command (addressed by 4ea7b2fea3332eaa7301569deeaa50a896d22f77).
Once that was fixed though we discovered that the command simply can't
be unmapped.  Any attempt to do so will result in an exception being
thrown.

This command is mostly a do-nothing command in 2010 so short term we
just translate it into <C-;> directly without the need to unmap.  Long
term we need to handle this in the same way that we handle Alt key
mappings.  Register our own command which has higher priority.

closes #967