Fixed gt and gT on recent builds
At some point there was a break down in the conditional compilation
logic I use for the Dev10 and Dev11 DLLs.  This caused the Dev10 layer
to identify as Dev11 and hence broke gt and gT on 2010.

Ultamitely this conditional compilation solution won't cut it anymore.
I need it to be more robust and sohuld factor it out into 2 DLLs that
don't use conditional compilation tricks.  It's just too easy to get
wrong in the msbuild file.

closes #975