More paragraph test cases
In trying to understand the behavior displayed in issue #978 I added
some more test cases for the paragraph motion.  It's still unclear why
the delete backwards is taking out the new line information in this case
so the tests are still failing.  Trying to find some documentation about
why they extend forward in this case.