Deploy script runs unit tests and verifies VSIX
The deploy script now runs the unit test DLLs in release and verifies
that the VSIX contents are correct.  The VSIX in particular is
troublesome because it's easy to mess up the contents and have files
randomly get added or deleted.

Can't fully verify the script because I have failing tests.  Will clean
them up with another change.