GetWpfTextView cleanup
Fixed a couple of places where calls to GetWpfTextView didn't account
for it throwing exceptions.  This can happen when the IVsTextView
instance isn't actually an editor shim.  This isn't of conesquence to
our code, we just want the IWpfTextView if it's available, how the
IVsTextView is implemented isn't a concern.