F12 doesn't put VsVim into Visual Mode
The last attempt to fix this missed the case that F12 can cause a new
file to be selected.  Unfortunately even though a new file is selected
it doesn't register as the active document until the next message pump.
The selection does occur synchronously though so we find the new
selections and clear them out to ensure we don't go into Visual Mode

closes #989