Click To Def on preview file caused IDE crash
When a click to definition was performed in a file opened in the preview
pane it would lead to a crash in VsVim.  The selection tracking code was
posting a delegate to deal with the selection change at a later point.
Before that delegate ran the IVimBuffer instance was closed.  When the
delegate later ran it was operating on a dead IVimBuffer and lead to a
crash