NavigateTo no longer enters Visual Mode
The C++ implementation of NavigateTo / QuickSearch was selecting the
target item as a part of navigation.  This was causing VsVim to
incorrectly enter Visual Mode.

To work around this I provided my own INavigateToItem implementation.
It looks for search starts and stops.  It won't allow the transition
into Visual Mode during the search and subsequent navigation operation

closes #1001