Released 1.3.2.0
Updated the version number to 1.3.2 to represent the new release.

Also updated the Deploy.ps1 script to deal with the new Colors.pkgdef
file.  The script closesly monitors the number of files in the resulting
VSIX and this new file triggered an error.  The file is expected though
so the script was adjusted to account for it.