Removed UseEditorSettings
The UseEditorSettings option was an ancient option that I added when my
support of Vim settings and synchronization with the editor was
extremely poor.  It was a way to opt out of my bad behavior.

Now the settings support is much better.  Instead of this option we use
the presence or abscence of the VimRC file to determine if VsVim or
Visual Studio settings should win