Fix unhandled COMException
During a navigation resulting from a NavigateTo UI, the target
ITextBuffer can fail to provide the set of ITextViews for the buffer.
This was resulting in an unhandled exception.

The point of the method is to return the views if available.  If they're
not available we should just be returning an empty collection.

closes #1013