No longer beeps on first Left arrow
The command line editting was beeping the first time the <Left> arrow
was hit.  The <Left> key was making it's way down to the Vim layer which
didn't understand how to handle <Left> in command mode and hence beeped.

The fix was to have the CommandMarginControllor intercept the <Left>
event and handle the key itself.