Fix "r" bug in HTML pages
This fixes the exception that occured when using 'r' to change tag names
in HTML pages.  The HTML document code is doing compensating edits to
keep the tag names in sync.  This happens after the edit by 'r' and
hence the buffer version was ahead of where VsVim expected it to be.

To fix this we no longer use the ITextSnapshot returned by Replace.
Instead query the ITextBuffer for the current ITextSnapshot.  This value
is then used to set the caret position.

closes #1040