Fix small delete register bug
The small delete register was being incorrectly updated.  It was being
updated for deletes that targetted a named register instead of deletes
that targetted the unnamed register.  That behavior was backwards and I
had 2 tests checking that backwards behavior.  Those tests were updated
and new integration tests were added for the scenario

closes #1037