Word motion on blank line was causing exception
The word forward motion on a blank line was causing an unhandled
exception to be thrown.  The logic that accounted for ending on a new
line and backing up over blank lines wasn't accounting for the case that
the caret line was itself blank.  Instead it backed up over it and tried
to create a SnapshotSpan with the points out of order (end before
start).

Fixed it to never back up over the caret line