Fixed a bug in the memory leak tests
It turns out a major source of the unreliablity was the
IKeyBindingService implementation.  It was looking for commands in the
created IVimBuffer *after* it was closed by the test method.  This
caused it to add event handlers after the fact and hence caused the
hooks to be added back in which were already expected to be freed.