Fix R# edits when both R# and dotCover installed on 2012
The VsVim behavior to respect R# live templates was broken in 2012 when
both R# and dotCover were installed.  The code didn't recognize the new
assembly names for R# in 2012, defaulted to the less correct logic which
was picking dotCover over R#.   Fixed to use the new assembly names

This is a solution to issue #1024. Unit tests still need to be written
though.  Also should make the backup logic stronger by prefering an
assembly with Resharper in the name over dotCover.