Refactored RegisterValue
The bugs I found on yanking and pasting the last line in the buffer made
it clear that the strings put into RegisterValue need to undergo some
sort of normalization process.  As a discriminated union it wasn't
possible to do this hence I converted to a type with a normal
constructor.  This will be our choke point for value normalization