VsVimShared now compiling
Had to do an awful hack for one of the XAML files here.  The
ConflictingKeyBindingDialog depends on a shell window.  The reference in
XAML includes the actual assembly name which changes between versions of
Visual Studio.   XAML as far as I know doesn't support #ifdef hence I
had to have 2 copies of the same file.  Hope I can get rid of this hack
at some point in the future.  It's pretty terrible