Vs2012 now compiling
This particular project is tied to a specific version of Visual Studio
hence it's a very straight forward change.  The version numbers can be
hard coded here.

The one exception is FSharp.Core.dll.  This DLL we want to force to the
same version as VimCore hence we use the standard version number