Fix key handling in Report Designer
The Report Designer hosts the WPF editor control and customizes the
processing of key strokes in a special way.  Based on the behavior it
looks like they are trying to get common combinations like Ctrl+A to
work in the hosted editor.  By default the hosting logic doesn't provide
this behavior and most common hosts have to manually add it.

They do this by overriding Control.IsInputKey.  If they provided key
stroke is one of the special keys then they pass it through
IOleCommandTarget::Exec.  If it is handled there then they return "true"
from IsInputKey.

This means they key is not considered handled and it will continue
through the WinForms chain of messages and eventually make it's way back
to any KeyProcessor implementations.  VsVim has both an
IOleCommandTarget and KeyProcessor hence it sees this key twice.

The fix is to just ignore all special keys in the KeyProcessor
implementation if we are dealing with the Report Expression Editor.

closes #1088