Fix bad removal of Backspace and Enter bindings
The 1.4.0 release would incorrectly strip off the binding for Enter and
Backspace when the Options dialog was run.  The was a bug in the diffing
code which make it believe that VsVim wanted to handle that key stroke.

The core problem was it asked the snapshot if Vim used the command
instead of asking the KeyBindingData if it wanted to do so.  The
KeyBindingData takes into account more factors like the fact that VsVim
never wants to control enter.  The snapshot data just says that VsVim is
indeed capable of handling enter.  I moved all of the logic back to
looking at the KeyBindingData and made the diffing processes much
cleaner.

Additionally I changed the code to fix this problem on the first startup
of the next release.

close #1101