Make ParseError a LineCommand value
Representing a parse error as a separate type was simply a bad
representation.  It made it impossible to mix parse errors into places
where a LineCommand value was desired.  For example in function bodies.

The core methods in Parser now return LineCommand instead of
ParseResult<LineCommand>.  Any routine wishing to know if there was a
parse error can simply query the LineCommand.ParseError union value.
This resulted in a major simplification of the parser.