Unit tests for issue 1136
This adds the basic unit tests for #1136.  The core problem appears to
be that we are tracking the repeated block edits as normal edits and not
a different command.  Hence it repeats as a simple edit instead of a
block edit.  Likely need a new InsertCommand for this to work