Exception in options caused VS crash
User reported an issue where hitting Ok on the options dialog caused
Visual Studio to crash.  While I was unable to reproduce this it's
pretty clear that the UpdateKeyBindings method was throwing an
exception.  It deals with a lot of COM objects, it's apparently capable
of throwing and sits on the core message loop.  Wrapped it with a try /
catch and reported the issue to the user