Added new key event
Added the KeyInputProcessing event.  This differs from KeyInputStarted
in that it will have the actual key which is being processed.  Hence
consumers can see the result of key maps and intercept the mapped key
before it is processed by vim.  KeyInputStarted only allowed consumers
to intercept keys which were typed.  A much less useful event