Failing to register normal as the previous mode
A bug in the change 31d9222 caused "normal" to fail to register as the
previous mode when transitioning to any "visual" mode.  This meant that
a subsequent leave of "visual" mode via <Esc> would go to insert, or
command line, instead of normal mode.  This fixes that problem and adds
some regression tests