Hookup 'scrolloff' to core movement commands
This change hooks up the scrolloff logic to the core movement commands.
This means they will all respect 'scrolloff' after changing the view.
There are likely several commands that I missed in this change as not
every command uses this APIs.  However it's largely working and I want
to dogfood it a bit while fixing up the other cases

More progress on #403