First round of overload removal
With the support of 'scrolloff' the types of operations involved with
moving the caret / positioning the view simply has too many combinations
to support via overloads (expand text, visible caret, scrolloff, virtual
space).  Hence we are moving to a simpler API + an enumeration to
specify the necessary operations.  This is the first change along those
lines which fixes everything but the virtual space

Related to #403