Fix crash in Navigate To hook
There is a bug in the RunningDocumentTable implementation in Vs2012 and
earlier that causes it to thrown an exception in the
IEnumerable<RunningDocumentInfo> implementation.  The code inadvertently
passes IntPtr.Zero to Marshal.Release (via GetDocumentInfo).  This
causes an exception which even if handled prevents enumeration of valid
documents.

This is fixed in Vs2013.  VsVim is still supported on earlier releases
so we have to manually enumerate via IVsRunningDocumentTable instead

Also added an exception handler into the code which initiated this call.
It runs directly on top of the message loop and hence unhandled
exceptions terminate the process.