Fixed build on Vs2010 only installation
The $(VisualStudioVersion) property can be empty on a machine with only
Visual Studio 2010 SP1 + the SDK installed.  This broke a couple of
project files which expected it to have the value 10.0 in this scenario.
Updated those projects to account for the missing value.

Also added a DLL that was needed to build without a 2012 installation on
the machine