Correct hookup of IOleCommandTarget with JavaScript
The JavaScript language service asynchronously hooks up its
IOleCommandTarget implementation.  It also incorrectly handles certain
keys.   For example it claims to handle <Esc> in QueryStatus but doesn't
process it in Exec.  This forces the VsVim implementation to be ahead of
it to be correct.

This change causes us to also asynchronously hookup IOleCommandTarget in
the presence of javascript.  It is not nearly reliable as I would like
it to be.  However it's sufficient for a quick patch to help with the
ReSharper issues i'm seeing.  Long term I'm going to be looking for a
better solution