Fix bad handling of '.' in VAX
There is a timing issue with VAX where for a short time after leaving
insert mode it will interpret '.' as the literal typing of a '.' vs. a
key stroke that should be handled by the editor (and consequently passed
to VsVim).  To work around this I special case the '.' key in
PreviewKeyDown to ensure that VsVim gets a chance to handle that key
when in normal mode.

The most reliable way to reproduce the original bug is to do the
following.  Start with a local that has an existing member access
expression

local.Test();

Put the caret on the T and execute the following key combination from
normal mode: "Cfield;<Esc>."  The time between <Esc> and '.' must be
very small (less than 2 seconds in my experimentation).  Instead of
repeating it will often just insert '.'.