Fix exception in PasteAdornment
The paste adornment implementation was querying certain properties of
the caret.  These properties all throw if the caret is off of the
screen.  This was resulting in an unhandled exception in certain
scenarios.  In particular if you did <C-R> from insert and then scrolled
the caret off the screen.  Simply catch the exception if it happens and
note that the adornment is not displayed

closes #1206