Fix Ctrl-Y and Ctrl-E to not scroll the caret
The Ctrl-Y and Ctrl-E commands should not be scrolling the caret but
instead scrolling the window.  The only time they should move the caret
is when the window scroll causes the caret to become invisible.  If that
occurs then the caret should be moved back into the visible portion of
the ITextView (while maintaining the caret column)

closes #1202