Moved R# key handling to a KeyProcessor
This change moves the special case key processing for R# into a
KeyProcessor.  This allows us to handle the special keys in the
PreviewKeyDown event.

This allows for much more definitive processing than IOleCommandTarget.
R# implements it's own IOleCommandTarget and we end up competing for key
strokes.  The VsVim design is to be resilient to missed / duplicate keys
but missing keys can lead to undesired behavior (such as double Esc to
leave insert when intellisense is up).  Moving to a KeyProcessor is more
reliable because R# at this time doesn't export one.  So I can intercept
the keys much earlier on in the process.

Of course anyone else could implement their own KeyProcessor and jump me
in line.  It's unlikely that R# would do this though and at this time
they don't which makes this more reliable