Fixed the failing unit tests
The VsKeyProcessorSimulation still wasn't entirely accurate.  When a key
down event is translated to an executed IOleCommandTarget event it
wasn't running the key up events.  This does not reflect reality.  The
OLE processing is done on key down only and is not a part of key up.
Changed it to reflect key up semantics

Also changed the R# KeyProcessor to respect discarded KeyInput values.
Unit tests are passing again