Removed the KeyDown handling
All of the commands we are trying to intercept are bound to VS commands.
This means they are handled in PreProcessMessage and never even make it
to the "down" side of key events.  This code was never meaningfully hit.

There is some up side to looking at key "up" events though.  VS doesn't
handle these in it's pre-processing layer and hence we would see them.
Future changes may do this