Show VsVim in the peek definition window
This change causes VsVim to show up in the peek definition window.  The
primary change here was switching the targeted text view role from
Document to Editable.  This is the more universal tag for vim support so
this makes sense.

This did cause us to initially pop up in interactive windows though.
Some instances define both Interactive and Editable like the NuGet
package manager console.  This is correct and we must handle it by
explicitly keeping ourselves out of interactive windows

Still some more work to do.  Escape in normal mode should cause the peek
window to be dismissed but we appear to be swallowing it

#1207