Cleaned up some equality implementations
The core equality tests weren't actually testing properly for null in
IEquatable<T>.  Changed it to properly test for that and fix up the
places which weren't correctly doing it.

Also implemented IEquatable on SearchData.  This is used within several
discriminated unions which require equality.  This needs to be done by
value not by reference