Refactored SearchService
This moved SearchService from a type with essentially only static
methods to one with instance methods.  Originally the static methods
were used to make it harder to use mutable state in unsafe ways.  This
is a type designed to be used from multiple threads and I wanted to make
it hard to get at data incorrectly.

Unfortunately this was also making it very hard to make progress on the
type.  Caching results for instance was really difficult.  Hence I made
the move to instance methods again.  The single piece of unsafe state
(_vimRegexOptions) is cached at every API entry point and passed along
the method calls