Added caching support to ISearchService
It's very common in vim to search for the same term multiple times.
This comes from simple user interaction (start a search, backspace a
typo, then search forward again) and in features keying off the same
data.  Searching is expensive especially in large files hence it makes
sense to cache the results in a simple fashion.  This introduces a very
simple 10 entry wide cache for search terms.