Fix parser hang on syntax error
In the case where a line of a script contained a single quote character
and nothing else the parser would enter an infinite loop due to 2 bugs.

The first was that ParseLineRangeSpecifier didn't properly unwind when
it found an incomplete line range.  That function should parse a line
range or not modify the token stream.  In this particular case it was
consuming the single quote and failing.  It now properly unwinds

The second bug is that in the parser didn't always move to the next
line.  This could only be exposed with the first bug but it was fixed
anyways to always move to a new line