Remove the old IKeyUtil implementation
The KeyUtil class was the original implementation of IKeyUtil.  This
implementation attempted to actually read and process keyboard layouts
which it turns out is an incredibly hard task.  It's also simply the
wrong way to approach the problem of key input mapping.  Textual input
is already done by WPF and that can be hooked into directly instead of
doing it (incorrectly) myself.  There are a subset of key strokes that
are meant to be mapped by virtual key and that is being done correctly
at this time.