Fixed the failing MemoryLeakTests suite
The MemoryLeak suite was failing due to the FallbackKeyProcessor
changes.  The problem was the handling of
IVimApplicationSettings.SettingsChanged event was never removed.  Since
IVimApplicationSettings has a lifetime equal to that of the host this
caused the FallbackKeyProcessor and all of its fields to be leaked.  The
fix is to remove the handler once the relevant view is closed.  At that
point key processing is no longer necessary