Moved undo down to IVimTextBuffer
Undo rightfully belongs in the buffer layer not the view.   This is how
the core editor associates undo and we were simply creating problems by
associating it with the ITextView instead.  This change moves undo down
to the IVimTextBuffer which is the logical pair for ITextBuffer.  Now
all IVimBuffer instances over the same ITextBuffer share the same undo
information.