Space now works in peek window
The space key now works properly in the peek window.  However there are
still some selection issues.

When the peek window is open there are actually 2 ITextViews for the
same ITextBuffer.  This happens because VS will open the peeked document
as a separate entry.  The selection is being mirrored by VS or VsVim,
unclear which, but not cleared properly.  This leads to the hidden
buffer having a selection but not the active buffer but VsVim keeps
believing there is a selection and hence keeps pushing the peek window
into Visual Mode.

related to #1207