Add parse support for :call
This adds parse support for the :call command but not actual functional
support.  A good number of vimrc files map keys to :call instructions.
The current implementation was giving almost no help to users because
that caused the key presses to result in "parse error".  This is simply
not actionable by users.

With this change the error message is very explicit and leads users to
realize that :call just isn't supported at this time.

#closes 1296